requirements.txt
asgiref==3.8.1
Django==4.2.13
django-autocomplete-light==3.11.0
pillow==10.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
